(gen): fix leaf types aggregation (#49)
Currently, in the leaf type aggregation, some intermediate types might end up in the result set because RMF creates subtypes on usages in properties. That is also the reason for the current names filter but it is not sufficient in all cases.
This can lead to unreachable code in the decoders.

With this change, we make sure the types we iterate through are API-level types. As a consequence, we might have some types that were previously intermediate types because they have (only) property subtypes. Those are considered leaf types now if the filtered subtypes set is empty.

+ update raml-model and related fixes