Improve support for Scala 3 code generation through Scalameta dialects
Also remove the hack to uppercase enums in #62, as that's not needed anymore.

Quasiquotes need the dialect to be fixed at compile time via import.
As it runs during compile time, we can't parameterize it with a dialect.
This is something we can't do here, as we want to dynamically toggle
the output variant, depending on the scala version of the project using
the plugin.

So to help keeping changes in check, we keep Scala 2 syntax in our quasiquote
literals, i.e. we parse q"import io.circe._"), and only apply the dialect during
pretty printing, which works for most things where the AST is the same.
I.e., imports have the same AST representation.